+++ services/payment_service/app/models/seller_wallet.py
@@ -10,7 +10,7 @@ class SellerWallet(Base):
 
     id: Mapped[uuid.UUID] = mapped_column(primary_key=True, default=uuid.uuid4)
     seller_id: Mapped[uuid.UUID] = mapped_column(nullable=False, unique=True, index=True)
-    initial_amount: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    current_balance: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    initial_amount: Mapped[float] = mapped_column(Integer, nullable=False, default=0)
+    current_balance: Mapped[float] = mapped_column(Integer, nullable=False, default=0)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=func.now())
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=func.now(), onupdate=func.now())
